+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Bug #1176: Index the result of `AbstractSchema::getSchemaMetadata()` by table name, so `getSchemaChecks()`,
   `getSchemaDefaultValues()`, `getSchemaForeignKeys()`, `getSchemaIndexes()`, `getSchemaPrimaryKeys()`,
   `getSchemaUniques()` and `getTableSchemas()` expose the table each item belongs to (@KalimeroMK)
+- New #1152: Add `UuidValue` expression that represents a UUID value independently of DBMS (@KalimeroMK)
 
 ## 2.0.1 February 09, 2026